Added Inventory and Variables Files sections to README
The README has been updated with two new sections: "Inventory" and "Variables Files". The "Inventory" section explains the need for an inventory file in Ansible, as it doesn't use an agent. It also includes a visual representation of an inventory file. The "Variables Files" section outlines best practices for storing host and group specific variables in their own files, along with the appropriate directories for hosts and groups.